+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## v1.0.5.2 ()
 - Delete files using system´s recycle bin
+- Copy a save game slot to another (empty) one
 
 ## v1.0.5.1 (16.10.2022)
 - Fixed a bug which deleted backup files while moving old backups to new location at the first start of v1.0.5.0